[IR] Add AuxData field to Instruction
This unsigned integer makes use of the currently available padding bytes
(size increase of 4 bytes on 32-bit platforms) to allow mapping
instructions to an integer quickly. The primary motivation is to
replace hash maps of Instruction pointers with dense vectors, where the
index into the vector is stored inside the instruction itself.