[SandboxIR] Delete SandboxIR/SandboxVectorizer
These components are still experimental, have not recieved commit
traffic beyond basic maintenance since May, and it was decided that the
people at Google working on the project should focus on other
priorities. Given the lack of investment, delete the code for now. If
someone wants to revive this, they can revert this patch. For now,
delete it so that upstream does not have to bear the maintenance cost.